Refactored zip extraction code to extractors package
